pyproject.toml
[project]
name = "bookrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "litellm>=1.72.6",
    "msgspec>=0.19.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "sentence-transformers>=4.1.0",
    "transformers>=4.52.4",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "rich>=14.0.0",
]
